pyproject.toml
[project]
name = "samir-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "fastapi>=0.124.4",
    "huggingface>=0.0.1",
    "huggingface-hub>=1.2.3",
    "iprogress>=0.4",
    "openai>=2.11.0",
    "pdfplumber>=0.11.8",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
